Pefetto v48.1

+++ src/trace_processor/perfetto_sql/stdlib/android/dvfs.sql
@@ -20,7 +20,7 @@ CREATE PERFETTO VIEW android_dvfs_counters(
   -- Timestamp when counter value changed.
   ts INT,
   -- Counter value.
-  value INT,
+  value DOUBLE,
   -- Counter duration.
   dur INT
 ) AS
@@ -68,9 +68,9 @@ CREATE PERFETTO TABLE android_dvfs_counter_stats(
   -- Counter name on which all the other values are aggregated on.
   name STRING,
   -- Max of all counter values for the counter name.
-  max INT,
+  max DOUBLE,
   -- Min of all counter values for the counter name.
-  min INT,
+  min DOUBLE,
   -- Duration between the first and last counter value for the counter name.
   dur INT,
   -- Weighted avergate of all the counter values for the counter name.
@@ -92,11 +92,11 @@ CREATE PERFETTO VIEW android_dvfs_counter_residency(
   -- Counter name.
   name STRING,
   -- Counter value.
-  value INT,
+  value DOUBLE,
   -- Counter duration.
   dur INT,
   -- Counter duration as a percentage of total duration.
-  pct FLOAT
+  pct DOUBLE
 ) AS
 WITH
 total AS (
